--- src/main/resources/bungee.yml
@@ -1,4 +1,0 @@
-name: EGlowAddon-Proxy
-version: '${version}'
-main: me.mrgraycat.eglowaddon.EGlowAddon
-description: A plugin to communicate with eGlow on the backend server for proper glow colors.
